+++ src/blocks.js
@@ -1,0 +1,2 @@
+import './sharing-link/index.js';
+import './sharing-label/index.js';