+++ src/sharing-link/icons/index.js
@@ -1,0 +1,7 @@
+export * from './facebook';
+export * from './chain';
+export * from './mail';
+export * from './pinterest';
+export * from './twitter';
+export * from './whatsapp';
+//# sourceMappingURL=index.js.map